#3: Added XML EDI orders reading and mapping
