+++ .gitignore
@@ -20,6 +20,7 @@ de.metas.endcustomer.mf15/metasfresh.properties
 **/.~lock*
 *.jasper
 *.log
+*.gz
 
 /de.metas.elasticsearch/data/
