+++ src/main/resources/application.properties
@@ -4,8 +4,8 @@ server.port=8184
 
 #rabbitmq config
 camel.component.rabbitmq.hostname=localhost
-#camel.component.rabbitmq.port-number=30050
-camel.component.rabbitmq.port-number=5672
+camel.component.rabbitmq.port-number=30050
+#camel.component.rabbitmq.port-number=5672
 camel.component.rabbitmq.username=guest
 camel.component.rabbitmq.password=guest
 
@@ -47,10 +47,8 @@ amqp.in.exchange.nondurable=de.metas.esb.from.metasfresh.nondurable
 # !! Deprecated !!
 # Fallback for older bundles. Every bundle should exlicitly use either the durable or the nondurable exchange!!
 amqp.in.exchange={{amqp.in.exchange.durable}}
-
-
-#edi.local.dir={{java.io.tmpdir}}/edi-standalone
-edi.local.dir=/home/esb/edi
+edi.local.dir={{java.io.tmpdir}}/edi-standalone
+#edi.local.dir=/home/esb/edi
 #
 # EDI Error Logging EndPoint Configuration
 #
@@ -184,6 +182,10 @@ edi.smooks.config.xml.orders.T100=/marshal-config/smooks/orders/edi-smooks-compu
 # reading from local file; see http://camel.apache.org/ftp.html on how to get the EDI data via FTP
 edi.input=file://{{edi.local.dir}}/input
 edi.input.orders={{edi.input}}/orders?preMove=.processed&move=.done&moveFailed=.error
+#
+# Input XML-to-XML
+#
+xml.edi.input.orders={{edi.input}}/orders/xml?preMove=.processed&move=.done&moveFailed=.error
 
 #
 # Feedback CHARSET_NAME
